create redux folder put app and feature in it and  rename file
