+++ client/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import App from './App';
-import store from './app/store';
+import store from './redux/app/store';
 
 ReactDOM.render(
   <Provider store={store}>
