+++ client/src/redux/app/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import useReducer from '../feature/userReducer';
+import useReducer from '../feature/user/userSlice';
 
 const store = configureStore({
   reducer: {
